[qca-nss-dp] Fix nss_dp_vp_skb_list len type from uint16_t to uint32_t
struct nss_dp_vp_skb_list 'len' field is uint16_t which can easily
wrap around for current EDMA budget. This patch changes the type from
uint16_t to uint32_t

+++ exports/nss_dp_vp.h
@@ -52,7 +52,7 @@ struct nss_dp_vp_rx_info {
 struct nss_dp_vp_skb_list{
 	struct nss_dp_vp_skb_list *next;
 	struct sk_buff_head skb_list;	/* skb list*/
-	uint16_t len;			/* Total data length carried by these skb*/
+	uint32_t len;			/* Total data length carried by these skb*/
 	uint8_t dvp;			/* Destination VP */
 };
